Fix bug in smu_get_dram_base_address() that improperly initialized command args for some processors.
smu_init_args() was completely ignored on the first command execution following
the switch statement used to determine the required command ID due to the goto
statement.

This led to an incorrect initialization of the args structure when for Matisse,
Vermeer and CastlePeak processors. The third, fourth and fifth arguments were
instead initialized to random values on the stack.

There were no known ill effects of this bug as the command itself only seems to
read the first 1 or 2 arguments in the argument address. However to be safe,
properly initialize the arg structure to zero.