Fixed PBO Scalar calculation.
The PBO scalar modification command used an incorrect calculation.
Whereas a 1X scalar translated to a 100 value and 10X scalar translated
to 1000 value, the reverse-engineering was incorrect in that this was
in fact the opposite.

This was confirmed by looking at some BIOS code where each scalar is in
fact multiplied by 100, where the default value is 100 under normal
operation.

Correct this documentation bug in the README example and rsmu_commands.md