Add Global Frequency Limit
CCLK_LIMIT gives the maximum frequency any processor core is
currently allowed to run at based on infrastructure limits
imposed by the hardware.

If boosting is disabled, this value has no meaning and defaults
to the maximum boostable frequency the processor is rated for.

It is common to see this value drop under workloads. The heavier
the workload, the more this infrastructure limit decreases, most
notably during heavy AVX instructions.